Move sync schema and related consts to separated folder
