+++ workers/loc.api/sync/schema/const.js
@@ -2,8 +2,10 @@
 
 const CONSTR_FIELD_NAME = '__constraints__'
 const TRIGGER_FIELD_NAME = '__triggers__'
+const ID_PRIMARY_KEY = 'INTEGER NOT NULL PRIMARY KEY AUTOINCREMENT'
 
 module.exports = {
   CONSTR_FIELD_NAME,
-  TRIGGER_FIELD_NAME
+  TRIGGER_FIELD_NAME,
+  ID_PRIMARY_KEY
 }
